Move get_grid_options() in parameter.py to grid_optimizer.py
