+++ tests/test_data/config_grid.json
@@ -33,6 +33,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -41,6 +42,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -49,6 +51,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -57,6 +60,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -65,6 +69,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -73,6 +78,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -81,6 +87,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -89,6 +96,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -97,6 +105,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       }
