+++ tests/test_data/config_grid_resumption.json
@@ -33,6 +33,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -41,6 +42,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -49,6 +51,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -57,6 +60,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -65,6 +69,7 @@
         "type": "uniform_float",
         "step": 1.0,
         "log": false,
+        "base": 10,
         "lower": -5.0,
         "upper": 5.0
       },
@@ -72,6 +77,7 @@
         "name": "x7",
         "type": "uniform_float",
         "step": 1.0,
+        "base": 10,
         "log": false,
         "lower": -5.0,
         "upper": 5.0
@@ -80,6 +86,7 @@
         "name": "x8",
         "type": "uniform_float",
         "step": 1.0,
+        "base": 10,
         "log": false,
         "lower": -5.0,
         "upper": 5.0
@@ -88,6 +95,7 @@
         "name": "x9",
         "type": "uniform_float",
         "step": 1.0,
+        "base": 10,
         "log": false,
         "lower": -5.0,
         "upper": 5.0
@@ -96,6 +104,7 @@
         "name": "x10",
         "type": "uniform_float",
         "step": 1.0,
+        "base": 10,
         "log": false,
         "lower": -5.0,
         "upper": 5.0
